style: textStyle and renaming playerhealthbar
